requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Google Generative AI integration
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Vector store
langchain-chroma>=0.1.0
chromadb>=0.4.0

# Document loaders and processing
pypdf>=3.17.0

# Environment variables
python-dotenv>=1.0.0

# Data processing (for judge scripts)
pandas>=2.0.0

# OCR Dependencies
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
